+++ Clinica/Models/ListarProfesionalesViewModel.cs
@@ -11,9 +11,9 @@ namespace Clinica.Models
         public string Nombre { get; set; }
         public string Apellido { get; set; }
         public string TipoDocumento { get; set; }
-        public decimal NumeroDocumento { get; set; }
+        public int NumeroDocumento { get; set; }
         public string Direccion { get; set; }
-        public decimal Telefono { get; set; }
+        public int Telefono { get; set; }
         public string Mail { get; set; }
         public string Sexo { get; set; }
         public int Matricula { get; set; }
